Refactor exception package
