+++ src/main/java/tecna/Tecna.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import org.json.simple.parser.ParseException;
 
 import tecna.collection.TaskList;
+import tecna.exception.JsonLoadingException;
 import tecna.storage.Storage;
 import tecna.ui.Ui;
