+++ src/main/java/tecna/parser/EventParser.java
@@ -5,8 +5,8 @@ import java.time.format.DateTimeFormatter;
 
 import org.json.simple.JSONObject;
 
-import tecna.JsonLoadingException;
-import tecna.JsonLoadingExceptionType;
+import tecna.exception.JsonLoadingException;
+import tecna.exception.JsonLoadingExceptionType;
 import tecna.task.Event;
 
 /**
